Add files via upload
As we know, globalization and technology are changing the world we live in. This is producing data at an unprecedented speed and volume, generating an information overload. For this reason, innovative solutions that process and analyze large amounts of data are a need for companies striving to make the most profit and be at the top of their game.
But is it just analytics, running and deploying machine learning models? What do we do with all the data collected? How is it stored? How can we access it? How can we know what type of information we have in a fast way? This information overload has raised new opportunities in many different sectors and an interesting solution raised: Text summarization.
Despite not being a brand-new solution, text Summarization is a very relevant topic in recent years due to better performance of models in the way they summarized texts. In the educational sector, where time is a constraint, and knowledge is produced by the second, a challenging scenario is in front of students. Especially on higher education, where work overload for MBA programs is extenuating, and the number of resources to be consumed are infinite.
In a small survey we conducted on 20 former MBA students at IE, we discovered that they were assigned an average of 120 business cases throughout the program. In such cases, nearly 60 % of them were read by the surveyed students. When asked about the reason they desist to read the other business cases, 80% of them answered that they were too long and time-consuming.
With this reality so close to us, it wouldn’t be time-efficient to have an overview of a long paper before deep diving on it? What about a brief summary of a complex business case? With that in mind, we created and present to you our Text Summarizer: Read Smarter, Learn Faster.
The project covers an application that allows students of Undergraduate and Graduate Programs at IE to summarize online sources and business cases, efficiently saving time in the research and studying stages. In order to accomplish this goal, we have analyzed and developed a set of text summarizers: 4 extractive and 1 abstractive text summarizers.
To be able to have solid models, we had to use proper data to evaluate our models. It was difficult to find business cases with their respective summaries that could be used. That is why we rather used a dataset provided by the BBC containing articles from 2004 to 2005 with their respective summaries. We specifically used the business-related articles to make sure our models were performant on business-related texts. Additionally, we tested our models on IE-provided cases to make sure their output was relevant. The metric used to assess the results of the models is Recall-Oriented Understudy for Gisting Evaluation (ROUGE), with a 3 granular output.
The best extractive summarizer is the Text Rank Model with scores of around 0.60 for R1, 0.47 for R2, and 0.57 for R-L. This model offered better readability when it comes to the output. It considers semantically similar words by using the embeddings approach. We here dive deeper into the theory behind this algorithm. The script of the various mentioned models is provided along with this report, we highly suggest spending a good amount of time reading it as most of our work can be understood from there as well.